pyproject.toml
[project]
name = "obsidian-import"
version = "0.1.0"
description = "Extract files (PDF, DOCX, PPTX, XLSX) into Obsidian-flavored Markdown"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Matthias Christenson" },
]
keywords = ["obsidian", "markdown", "pdf", "docx", "pptx", "xlsx", "import", "extraction"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Office/Business",
]
dependencies = [
    "pyyaml>=6.0,<7",
    "click>=8.0,<9",
    "pdfplumber>=0.11,<1",
    "pypdf>=5.0,<6",
    "defusedxml>=0.7,<1",
    "python-pptx>=1.0,<2",
    "openpyxl>=3.1,<4",
]

[project.optional-dependencies]
markitdown = ["markitdown[all]>=0.1"]
docling = ["docling>=2.0"]
dev = ["pytest>=8.0,<9", "hypothesis>=6.0,<7", "pytest-cov>=5.0,<6"]

[project.urls]
Documentation = "https://neuralsignal.github.io/obsidian-import/"
Repository = "https://github.com/neuralsignal/obsidian-import"

[project.scripts]
obsidian-import = "obsidian_import.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["obsidian_import"]

[tool.hatch.build.targets.sdist]
include = ["obsidian_import/"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501", "SIM108"]

[tool.pytest.ini_options]
testpaths = ["tests"]
